Show the completed percentage as each finding chart is downloaded
Let the user know of the percentage of the download of each finding
chart that has been completed. This is achieved by using the function
urllib.urlretrieve(), whose third argument defines the hook function
that is called once the network connection is established and every
time a new block is read. This allows us to display and keep updated
a message such as "Downloading M31.jpg: 87 %".

If any exception is raised by urllib.urlretrieve() (and this includes
KeyboardException, if the user stops the execution), the file being
downloaded is silently removed before the exception is re-raised.
This guarantees that we will not be left with truncated data.