Exit the script on KeyboardInterrupt
If urllib.urlretrieve() raises any exception, it is caught aynd an
error message informing the user that something went wrong, such as
"Download of NGC4321.jpg failed", is displayed and the script then
moves to the next image. However, the KeyboardInterrupt (raised by
Control-C) should stop the execution of the script, and not only of
the download of a particular finding chart.