Use multiprocessing module to download the finding charts in parallel
As suggested by Juan de Dios Santander, use a multiprocessing pool of
workers to download multiple finding charts simultaneously. The number
of workers in the pool is equal to that of CPUs in the system (e.g.,
if there are four cores on the machine, charts are downloaded from the
SDSS server in batches of eight).