Do not show completed percentages as finding charts are downloaded
Now that the finding charts are downloaded in parallel we cannot
display the percentage of each download that has been completed so
far, as multiple downloads would update the percentage at the same
time. We need to revert to the original output, showing a slightly
more detailed information so that the user always knows to which
finding chart each message refers.